Organize files and add required packages
