+++ .gitignore
@@ -17,3 +17,4 @@ work-record-archive
 export
 internal
 data
+.Rproj.user
